% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cetcolor-package.R
\docType{package}
\name{cetcolor-package}
\alias{cetcolor}
\alias{cetcolor-package}
\title{cetcolor: CET Perceptually Uniform Colour Maps}
\description{
Collection of perceptually uniform colour maps made by Peter Kovesi (2015) "Good Colour Maps: How to Design Them" \href{https://arxiv.org/abs/1509.03700}{arXiv:1509.03700} at the Centre for Exploration Targeting (CET).
}
\details{
56 Colour maps are available as of May 2018
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless-rpkg/cetcolor}
  \item \url{https://r-pkg.thecoatlessprofessor.com/cetcolor/}
  \item \url{http://peterkovesi.com/projects/colourmaps/}
  \item Report bugs at \url{https://github.com/coatless-rpkg/cetcolor/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Peter Kovesi \email{peter.kovesi@uwa.edu.au} (Creator of CET Color Maps) [copyright holder]
}

}
\keyword{internal}
